requirements.txt
mtcnn==0.1.0
tensorflow==2.0.0     
keras-vggface==0.6       
keras_applications==1.0.8
pyyaml
tqdm
scikit-learn
streamlit
bing-image-downloader


